+++ CHANGELOG.md
@@ -3,8 +3,8 @@ docxBox Changelog
 
 V0.0.2 - Not finished yet
 -------------------------
-* Add replace text by WML table
-
+* Add remove content between text (rem)
+* Add replace text by table (rpt)
 
 V0.0.1 - 2020-04-20
 -------------------